fix: split test bootstrap file
