+++ test/DvsaDocument/Service/Document/DocumentServiceTest.php
@@ -250,8 +250,10 @@ class DocumentServiceTest extends TestCase
             'repeatable' => ['1', '2']
         ];
 
-        // this is intentional
         /**
+         * This test intentionally passes an argument of the wrong type
+         * to the function and checks that it throws an exception
+         *
          * @psalm-suppress InvalidArgument
          * @phpstan-ignore-next-line
          */
